app/requirements.txt
customtkinter
pyserial
pillow